==> notebooks/1_bank_preprocessing.ipynb <==
import pandas as pd

bank_data_frame = pd.read_csv(r'C:\Users\vipki\Desktop\Python\preprocessing_and_vizualization\Data\Bank Marketing Data Set\bank.csv', sep=';')

bank_data_frame.head()

bank_data_frame.info()

bank_data_frame.isnull().sum()

bank_data_frame.mean()

bank_data_frame.dropna()

pd.get_dummies(bank_data_frame)

pd.get_dummies(bank_data_frame).to_csv(r'C:\Users\vipki\Desktop\Python\preprocessing_and_vizualization\Data\Bank Marketing Data Set\bank_preprocessed.csv', index_label = False)

df1 = pd.read_csv(r'C:\Users\vipki\Desktop\Python\preprocessing_and_vizualization\Data\Bank Marketing Data Set\bank_preprocessed.csv', sep=',')

df1
