==> notebooks/1_student-mat_preprocessing.ipynb <==
import pandas as pd

student_data_frame = pd.read_csv(r'C:\Users\vipki\Desktop\Python\preprocessing_and_vizualization\Data\Student Performance Data Set\student-mat.csv', sep=';')

student_data_frame.head()

student_data_frame.info()

student_data_frame.isnull().sum()

student_data_frame.mean()

student_data_frame.dropna()

pd.get_dummies(student_data_frame, prefix='school', columns=['school'])

pd.get_dummies(student_data_frame)

pd.get_dummies(student_data_frame).to_csv(r'C:\Users\vipki\Desktop\Python\preprocessing_and_vizualization\Data\Student Performance Data Set\student-mat_preprocessed.csv', index_label = False)

df1 = pd.read_csv(r'C:\Users\vipki\Desktop\Python\preprocessing_and_vizualization\Data\Student Performance Data Set\student-mat_preprocessed.csv', sep=',')

df1.head()
