==> notebooks/2_bank_classification.ipynb <==
import pandas as pd

dataset = pd.read_csv(r'C:\Users\vipki\Desktop\Python\preprocessing_and_vizualization\Data\Bank Marketing Data Set\bank_preprocessed.csv', sep=',')

dataset.head()

dataset.info()

dataset["y_yes"].unique()

X = dataset.iloc[:, :-1].values
y = dataset.iloc[:, 52].values

print(y)
print(X)

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

Разбиваем выборку на ту, на которой будет производится обучение, и выборку для проверки качества

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

Скалирование выборок для уменьшения влияний больших значений относительно других параметров

X_train

# Метод k-ближайших соседей (K-Nearest Neighbors)

from sklearn.neighbors import KNeighborsClassifier
clf = KNeighborsClassifier(n_neighbors=5)
clf.fit(X_train, y_train)

from sklearn.metrics import classification_report, confusion_matrix

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

Из Confision матрицы можно сказать: 729 - True Positive, 3 - Fasle Positive
                                     57 - False Negative, 53 - True Negative
Верных срабатываний = 729, Ложные срабатывания = 3, Ложные пропуски = 57, Истинные пропуски = 53. Алгоритм срабатывает довольно точно, но страдает полнота в случае '1'.

# Классификатор дерева решений (Decision Tree Classifier)

from sklearn import tree

clf = tree.DecisionTreeClassifier()
clf = clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

Верных срабатываний = 795, Истинные пропуски = 110. Алгоритм срабатывает идеально.

# import graphviz
# dot_data = tree.export_graphviz(clf, out_file=None, filled=True, rounded=True, special_characters=True) 
# graph = graphviz.Source(dot_data)
# graph.render('s')
# graph = graphviz.Source(dot_data)  
# graph 

# Наивный байесовский классификатор (Naive Bayes)

from sklearn.naive_bayes import GaussianNB

clf = GaussianNB()
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

Верных срабатываний = 795, Истинные пропуски = 110. Алгоритм срабатывает идеально.

# Метод опорных векторов (Support Vector Machines)

from sklearn import svm
clf = svm.SVC()
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

Верных срабатываний = 795, Ложные срабатывания = 1, Ложные пропуски = 5, Истинные пропуски = 104. Алгоритм срабатывает довольно хорошо, близко к идеальному.

# Логистическая регрессия (Logistic Regression)

from sklearn.linear_model import LogisticRegression

clf = LogisticRegression()
clf.fit(X_train, y_train)

y_pred = clf.predict(X_test)
print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

Верных срабатываний = 795, Истинные пропуски = 110. Алгоритм срабатывает идеально.

# Разделение выборки

from collections import Counter
print(Counter(y_train).values())
print(Counter(y_test).values())

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20, stratify=y)
print(Counter(y_train).values())
print(Counter(y_test).values())

Разбиваем данные более корректно для улучшения работы алгоритмов (получения корректных результатов)

# Кросс-валидация

from sklearn.model_selection import StratifiedKFold

skf = StratifiedKFold(n_splits=5)
for train_index, test_index in skf.split(X, y):
  print("TRAIN:", train_index, "TEST:", test_index)
  X_train, X_test = X[train_index], X[test_index]
  y_train, y_test = y[train_index], y[test_index]

from sklearn.model_selection import StratifiedKFold

skf = StratifiedKFold(n_splits=5)
for train_index, test_index in skf.split(X, y):
  X_train, X_test = X[train_index], X[test_index]
  y_train, y_test = y[train_index], y[test_index]
  clf = KNeighborsClassifier(n_neighbors=5)
  clf.fit(X_train, y_train)
  y_pred = clf.predict(X_test)
  print(classification_report(y_test, y_pred))
  print(clf.score(X_test, y_test))

Качество выполнения алгоритма примерно одинаково на всех выборках

# Подбор гиперпараметров

from sklearn.model_selection import GridSearchCV

from sklearn.neighbors import KNeighborsClassifier

parameters = {'n_neighbors':[1, 2, 3, 4, 5]}

knn = KNeighborsClassifier()

clf = GridSearchCV(knn, parameters)

clf.fit(X_train, y_train)

clf.cv_results_

Видно, что наибольшую корректность нам даст выбор 5 соседей на данной выборке
