==> notebooks/4_ensemble_bank_marketing_classification.ipynb <==
import pandas as pd
from sklearn import metrics
from sklearn import ensemble  
from sklearn.model_selection import GridSearchCV
import numpy as np
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import BaggingClassifier

dataset = pd.read_csv(r'D:\Fork\data-analysis\Data\Bank Marketing Data Set\bank_preprocessed.csv', sep=',')

dataset.head()

X = dataset.iloc[:, :-1].values
y = dataset.iloc[:, 52].values

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

 Метод опорных векторов (Support Vector Machines)

from sklearn import svm
svm = svm.SVC()
svm.fit(X_train, y_train)

y_pred = svm.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

 Бэггинг

%%time
bg = ensemble.BaggingClassifier(svm, n_jobs = -1)
bg.fit(X_train, y_train)
y_pred = bg.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

bagging_parameters_grid = { 'n_estimators' : [1, 50, 100],
                           'max_features' : range(1,10,1)
    
}
bagging_grid = GridSearchCV(bg, bagging_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
bagging_grid.fit(X_train, y_train)
bagging_grid.best_params_, bagging_grid.best_score_

y_pred = bagging_grid.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

Бустинг

%%time
bs = ensemble.GradientBoostingClassifier()
bs.fit(X_train, y_train)
y_pred = bs.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

boosting_parameters_grid = { 'learning_rate' : np.arange(0.1, 0.8, 0.1),
                           'n_estimators' : [10, 50, 100],
                            'max_depth' : range(1,10,1)

}

boosting_grid = GridSearchCV(bs, boosting_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
boosting_grid.fit(X_train, y_train)
boosting_grid.best_params_, boosting_grid.best_score_

y_pred = boosting_grid.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

 Стекинг

estimators = [
('svm', SVC()),
('knn', KNeighborsClassifier(algorithm = 'brute',metric = 'manhattan', n_neighbors = 9))
]

%%time
sk = ensemble.StackingClassifier(estimators=estimators)
sk.fit(X_train, y_train)
y_pred = sk.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))

stacking_parameters_grid = { 'stack_method' : ['auto', 'pred_proba', 'solution_function', 'predict'],
                            'cv' : [1, 2, 3, 4, 5, 6, 7, 8, 9]
    
}

stacking_grid = GridSearchCV(sk, stacking_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
stacking_grid.fit(X_train, y_train)
stacking_grid.best_params_, stacking_grid.best_score_
y_pred = stacking_grid.predict(X_test)
print(metrics.confusion_matrix(y_test, y_pred))
print(metrics.classification_report(y_test, y_pred))
