==> notebooks/4_ensemble_student_perfomance_regression.ipynb <==
import numpy as np 
import pandas as pd 
import matplotlib.pyplot as plt
from sklearn.ensemble import BaggingRegressor
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import GradientBoostingRegressor

dataset = pd.read_csv(r'D:\Fork\data-analysis\Data\Student Performance Data Set\student-mat_preprocessed.csv', sep=',')

бэггинг

dataset.head()

X = dataset.iloc[:, :-1].values
y = dataset.iloc[:, 15].values

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)
# разбиваем выборку

scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)
# Скалирование выборок для уменьшения влияний больших значений относительно других параметров

model = BaggingRegressor().fit(X_train, y_train)
test_score=model.score(X_test, y_test)
train_score=model.score(X_train, y_train)
train_score

from sklearn.model_selection import cross_val_score
score=cross_val_score(model, X, y, cv=10)
score

bagging_reg_parameters_grid = { 'n_estimators' : [10, 50, 100],
                               'max_samples' : range(1, 5, 1),
                               'max_features' : range(1, 5, 1)
}

bagging_reg_grid = GridSearchCV(BaggingRegressor(), bagging_reg_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
bagging_reg_grid.fit(X_train, y_train)
bagging_reg_grid.best_params_, bagging_reg_grid.best_score_

бустинг

from sklearn.model_selection import KFold
from sklearn.ensemble import GradientBoostingClassifier

dataset = pd.read_csv(r'D:\Fork\data-analysis\Data\Student Performance Data Set\student-mat_preprocessed.csv', sep=',')

X = dataset.iloc[:, :-1].values
y = dataset.iloc[:, 15].values
num_tree = 100
max_features = 7

kfold = KFold(n_splits=10)
model = GradientBoostingClassifier(n_estimators=num_tree,random_state=seed)
result = cross_val_score(model, X, y, cv=kfold)
print("result:",result.mean())

dataset= pd.read_csv('student-mat_preprocessed.csv', sep=',')

X = dataset.iloc[:, :-1].values
y = dataset.iloc[:, 15].values

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)
# разбиваем выборку

scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)
# Скалирование выборок для уменьшения влияний больших значений относительно других параметров

boosting_reg_parameters_grid = { 'loss' : ['ls', 'lad', 'huber', 'quantile'],
                                'learning_rate' : np.arange(0.1, 0.8, 0.1),
                                'n_estimators' : [10, 50, 100 , 150]
}

boosting_reg_grid = GridSearchCV(GradientBoostingRegressor(), boosting_reg_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
boosting_reg_grid.fit(X_train, y_train)
boosting_reg_grid.best_params_, boosting_reg_grid.best_score_

Cтекинг

from sklearn.datasets import load_diabetes
from sklearn.linear_model import RidgeCV
from sklearn.svm import LinearSVR
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import StackingRegressor
from sklearn.tree import DecisionTreeRegressor

estimators = [('lr', RidgeCV()), ('svr', LinearSVR(random_state=42))]
reg = StackingRegressor(estimators=estimators)

model = reg.fit(X_train, y_train)

test_score=reg.score(X_test, y_test)
train_score = reg.score(X_train, y_train)
train_score

test_score

score=cross_val_score(model, X, y, cv=10)
score

stacking_reg_parameters_grid = {'cv' : [1, 2, 3, 4, 5, 6, 7, 8, 9]}

stacking_reg_grid = GridSearchCV(reg, stacking_reg_parameters_grid,
                        cv=5, n_jobs=-1, verbose=True)
stacking_reg_grid.fit(X_train, y_train)
stacking_reg_grid.best_params_, stacking_reg_grid.best_score_
