==> notebooks/5_neural_network.ipynb <==
import numpy as np
import pandas as pd
from sklearn.preprocessing import LabelEncoder
from sklearn import datasets
from sklearn.metrics import classification_report, confusion_matrix, r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler

import os
import sys
sys.path.append(os.path.abspath('../'))
from modules import mlp

# Датасет Titanic

## Подготовка данных

Импортируем датасет

data = pd.read_csv('../Data/titanic.csv', sep = ',')

data

data.drop(['PassengerId', 'Name', 'Ticket', 'Cabin'], axis=1, inplace=True)
data.fillna(data.mean(), inplace=True)
data

data = pd.get_dummies(data=data, prefix='c', columns=['Sex', 'Embarked'])
data

y = data[['Survived']]
X = data[data.columns[1:data.shape[1]]]
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y)

## Классификация с помощью модели MLP

model = mlp.MLP(inputs_count=X_train.shape[1], hidden_layers_counts_with_act_funcs=[(15, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative),
                                                                      (15, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative)])\
    .fit(X_train.values, y_train.values, epochs=600, learning_rate=0.25)

y_pred = []
for i in range(len(X_test.values)):
    prediction = model.predict(X_test.values[i])[0]
    if prediction > 0.5:
        y_pred.append(1)
    else:
        y_pred.append(0)

print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

# Датасет Iris

## Подготовка данных

 Импортируем датасет

iris = datasets.load_iris()

X = iris.data
X

y = iris.target
y

 X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, stratify=y)

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

##  Классификация с помощью модели MLP

model = mlp.MLP(inputs_count=X_train.shape[1], hidden_layers_counts_with_act_funcs=[(3, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative),
                                                                      (3, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative)],
                outputs_count_with_act_funcs=[(3, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative)])\
    .fit(X_train, y_train, epochs=300, learning_rate=0.15)

y_pred = []
for i in range(len(X_test)):
    y_pred.append(np.argmax(model.predict(X_test[i])))

 print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

# Bank Marketing Data Set

## Подготовка данных

 Импортируем датасет

bank_df = pd.read_csv('../Data/Bank Marketing Data Set/bank_preprocessed.csv', sep = ',')

bank_df

X = bank_df.iloc[:, :-1].values
print(X)

y = bank_df.iloc[:, 8].values
print(y)

Разбиваем исходные данные на обучающую и контрольную выборки. Обучающая выборка составляет 80% от объема исходных данных, а контрольная составляет, соответственно, 20%

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

Объявляем StandardScaler для нормализации данных

scaler = StandardScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)

## Модель MLP

model = mlp.MLP(inputs_count=X_train.shape[1], hidden_layers_counts_with_act_funcs=[(3, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative),
                                                                      (3, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative)],
                outputs_count_with_act_funcs=[(3, mlp.ActivationFunctions.sigmoid, mlp.ActivationFunctions.sigmoid_derivative)])\
    .fit(X_train, y_train, epochs=300, learning_rate=0.15)

y_pred = []
for i in range(len(X_test)):
    y_pred.append(np.argmax(model.predict(X_test[i])))

print(confusion_matrix(y_test, y_pred))
print(classification_report(y_test, y_pred))

## Выводы
 
Выполнена своя реализация многослойного персептрона. 
Модель протестирована на рекомендуемых датасетах. 
Качество классификации получилось заметно хуже, чем при использовании базовых моделей
