==> notebooks/7_kearas_classification.ipynb <==
import tensorflow as tf
from tensorflow import keras
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd 
from sklearn.model_selection import train_test_split

df = pd.read_csv("D:/Fork/data-analysis/Data/Bank Marketing Data Set/bank_preprocessed.csv")

Для начала преобразуем датасет: удалим 1 лишний дулбирующий столбец целевого признака

df.rename(columns={'y_yes': 'y'}, inplace=True)
df.drop('y_no', axis=1, inplace=True)

df.head()

df.info()

Разобьем выборку на обучающую и тестовую

X = df.iloc[:, :-1].values
y = df.iloc[:, 51].values
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

y_test

Создадим Sequential model №1: на 1 слое 10 нейронов, на 2 - 5, на 3 - 1 выходной нейрон, активационные функции: relu и сигмоида

# тест
model = keras.Sequential()
model.add(keras.layers.Dense(10, input_dim=len(X_train[0,:]), activation='relu'))
model.add(keras.layers.Dense(5, activation='relu'))
model.add(keras.layers.Dense(1, activation='sigmoid'))

model.summary()

model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])

X_train.shape

y_train.shape

model.fit(X_train, y_train, epochs=30, batch_size=128, shuffle=True)

test_loss, test_acc = model.evaluate(X_test, y_test)
print('Test accuracy:', test_acc)

Создадим Sequential model №2 с количествами нейронов в виде степеней 2, активационные функции: relu и сигмоида, также добавлены dropout для борьбы с переобучением

# тест 2
model1 = keras.Sequential()
model1.add(keras.layers.Dense(128, input_dim=len(X_train[0,:]), activation='relu'))
model1.add(keras.layers.Dropout(0.3))
model1.add(keras.layers.Dense(64, activation='relu'))
model1.add(keras.layers.Dropout(0.3))
model1.add(keras.layers.Dense(32, activation='relu'))
model1.add(keras.layers.Dropout(0.3))
model1.add(keras.layers.Dense(1, activation='sigmoid'))

model1.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])

model1.summary()

model1.fit(X_train, y_train, epochs=30, batch_size=128, shuffle=True)
# 1 модель норм, для модели не нужно много нейронов

test_loss, test_acc = model1.evaluate(X_test, y_test)
print('Test accuracy:', test_acc)

Для 3 модели взят случайный набор количества нейронов

# тест 3
model2 = keras.Sequential()
model2.add(keras.layers.Dense(20, input_dim=len(X_train[0,:]), activation='relu'))
model2.add(keras.layers.Dense(15, activation='relu'))
model2.add(keras.layers.Dense(10, activation='relu'))
model2.add(keras.layers.Dense(5, activation='relu'))
model2.add(keras.layers.Dense(1, activation='sigmoid'))

model2.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])

model2.fit(X_train, y_train, epochs=30, batch_size=128, shuffle=True)

test_loss, test_acc = model2.evaluate(X_test, y_test)
print('Test accuracy:', test_acc)

Исходя из оценки Точности моделей на тестовых выборках можно сделать вывод, что 1 модель справилась куда лучше других. Для решения задачи данного датасета с помощью нейроных сетей keras не требуется создавать нейронную сеть с большим количеством нейронов и слоев.

model.save("../keras_models/keras1.h5")
model1.save("../keras_models/keras2.h5")
model2.save("../keras_models/keras3.h5")
